alloc: Allow disabling memfd usage at runtime with XSHMFENCE_NO_MEMFD=1
For thin VMs used for compatibility reasons, it is desirable to directly
proxy the X11 protocol between an application running in a VM and an X
server running on the host. We can proxy futex wakeups with some watcher
code, but we need the futex memory itself to be shared. This works when
/dev/shm is mounted with virtiofs with DAX, but cannot work for memfds.

Add a XSHMFENCE_NO_MEMFD environment variable that, when set, disables
the memfd codepath and falls back to /dev/shm.

Signed-off-by: Asahi Lina <lina@asahilina.net>
Part-of: <https://gitlab.freedesktop.org/xorg/lib/libxshmfence/-/merge_requests/9>